Fix inclusive jump
This commit uses visual mode to fix inclusive jump. The change of the
`virtualedit` option is no longer required. The auto-detection of
operation-pending mode makes it possible to use the same config for all
modes. I checked the prefix of the mode string, but I'm not sure whether
the behavior is correct or not when "forced char/line/blockwise".

The `inclusive_jump` option must be enabled by default, because in
normal mode the default behavior should be inclusive jump.

The documentation (README.md and doc/hop.txt) is not modified. This is
also a breaking change which need to be carefully handled. I'll leave
these to the maintainers.